+++ swarm/CHANGELOG.md
@@ -29,6 +29,10 @@
   This was never constructed.
   See [PR 3373].
 
+- Remove `DialError::ConnectionIo` variant.
+  This was never constructed.
+  See [PR 3374].
+
 [PR 3364]: https://github.com/libp2p/rust-libp2p/pull/3364
 [PR 3170]: https://github.com/libp2p/rust-libp2p/pull/3170
 [PR 3134]: https://github.com/libp2p/rust-libp2p/pull/3134
@@ -39,6 +43,7 @@
 [PR 3188]: https://github.com/libp2p/rust-libp2p/pull/3188
 [PR 3377]: https://github.com/libp2p/rust-libp2p/pull/3377
 [PR 3373]: https://github.com/libp2p/rust-libp2p/pull/3373
+[PR 3374]: https://github.com/libp2p/rust-libp2p/pull/3374
 
 # 0.41.1